requirements.txt
flask
flask-socketio
gunicorn
fastapi
uvicorn
python-dotenv
pydantic